[serverless] Refactor layout package
